// src/modules/configuration/model/JsonService.ts
export interface JsonService {
    JsonServiceClassName: string;
    IDJsonService: number;
    ResponseJsonValue: string;
    Description: string;
    URLService: string;    
    Name: string;
    IDEmployee: string;
    IDJsonServiceClass: number;
    State: number;
    InputJsonValue: any;
    URLRequestType: any;
}